AD-HOC refactor (Dependencies): Remove geerlingguy.nginx dependency
When this module was extracted out of the project in which it was
initially written, it was a dependency of geerlingguy.ngnix to take
advantage of the ordering capabilities of Ansible.

However, with its' release it's used in contexts that are not simply
geerlinguy's roles. Accordingly, this dependency no longer accurately
reflects it's use, and instead it should be manually ordered in the
playbook.