AD-HOC feat (Configuration): Add support for TLS
Currently, there is a problem in which it's possible to eavsdrop a
connection during authentication and pick up the username and password
used to connect to that connection. That username and password can then
be used arbitrarily by the eavsdropper in future.

This commit introduces support for TLS, which wraps the connection in a
loving cocoon of encryption. This encryption prevents the nefarious
eavsdropper from reading those credentials, met instead with only the
inspired tunes of Rick Astley.

Or something.