AD-HOC refactor (*): Add support for apache
Recently, Sitewards expanded the server management tooling to include
apache/mod_php. This is fine, however it's not reasonable to run an
extra webserver on the machine simply because it's easier to reverse
proxy content to metrics.

This commit adds support for the apache webserver. Necessarily, this
means refactoring the role to be much more independent.

Note: A design flaw of this work is that apache does not have TLS
support. This is because the author currently has no mechanism to test
this.

== Design Notes ==

=== Lint Configuration ===

This commit also includes validation tooling for the authors lint setup.
This is described at the following URL:

  https://medium.com/@andrewhowdencom/i-successfully-fake-being-a-tidy-
    developer-with-this-one-weird-trick-okay-git-hooks-733573b1c679