Adhoc (config) Allow an external port to be defined
While using this role I stumbeled over a port conflict.
This was preventing ngnix from restarting. The port was not
claimed by either nginx or the node-exporter on only the
interface that is belonging to the ip address it is bound

I therefor implemented a solution where the internal and
external port can be different. This change is introducing
the port_public variable.

Attation: This change is not backwords compatible.

There was also a hint in the 9199c52 commit. That is removing
a comment in templates/etc/nginx/sites-enabled/template.conf
this comment was indicating my current headaches. But it was
removed without giving further details.