fix: set proxy server templates to listen to defined public port
The `port_public` option introduced in the following commit but never
used in template creation while it's really usful.
https://github.com/sitewards/ansible-role-prometheus-proxy/blob/adhoc_fix_port_clash/defaults/main.yml#L12-L15

The problem with current approach is we only can configure reverse proxy
to use the same port for public listening and internal reversing and
it's not possible to revese traffic to a different internal port if
needed.
The new approach can cover both possible cases. Reversing traffic to a
different port and in case port must be the same, we still able to
define the same port number in config file.